Disable some xterm key sequence enhancements.
Space and Enter and Backspace with modifier keys are rarely bound, so
having support for them leads to Readline inserting "leftover" chars
from the unbound key sequences.  This is how Readline/etc have always
worked with respect to unbound keys.

Since the goal of this sample is to show the basics of using Readline
on Windows, it shouldn't do this more sophisticated stuff.